Improve Windows compatibility, and small logging cleanup.
Bump to version 0.2.

Take most actions on keyDown, except release of CTRL keys (which has to happen in keyUp).
Use preventDefault to suppress the default Windows behavior of most hotkeys.

Add version to log, and log once at beginning and once at end so we will always get at
least a version, and can tell when we initialized successfully.

Move "use strict" declaration to the top so it applies to the whole file.